Modification and creation of controllers files
