+++ BackendFerreSys/src/routes/indexRoutes.js
@@ -7,15 +7,29 @@ const router = Router();//Creamos la instancia router.
 
 //IMPORTAMOS LAS FUNCIONES-CONTROLADORES
 
-import { insertCliente, insertProveedor, insertMarca, insertCategoria, insertArticulo, insertReciboMercancia, insertEncabezadoVenta, insertDetalleVenta, insertUsuario, insertPermiso, insertUsuarioPermiso,
-updateCliente,updateProveedor,updateMarca, updateCategoria, updateArticulo, updateUsuario, updatePermiso, updateUsuarioPermiso,
-deleteCliente, deleteProveedor, deleteCategoria, deleteArticulo, deleteUsuario, deletePermiso, deleteUsuarioPermiso} from '../controllers/crudController';
+import { getClientesError, getClienteByIdError, insertCliente, updateCliente, deleteCliente  }from '../controllers/ClientesCRUD';
 
-import { getClientesError, getProveedoresError, getMarcasError, getCategoriasError, getArticulosError,
-		 getReciboMercanciasError, getKardexError, getUsuariosError, getPermisosError, getUsuarioPermisosError }from '../controllers/crudController'; 
+import { getProveedoresError, getProveedorByIdError, insertProveedor, updateProveedor, deleteProveedor  }from '../controllers/proveedoresCRUD';
 
-import { getClienteByIdError, getProveedorByIdError, getMarcaByIdError, getCategoriaByIdError, getArticuloByIdError, 
-getReciboMercanciaByIdError, getKardexByIdError,getUsuarioByIdError, getPermisoByIdError, getUsuarioPermisoByIdError } from '../controllers/crudController'; 
+import { getMarcasError, getMarcaByIdError, insertMarca, updateMarca, deleteMarca }from '../controllers/marcasCRUD';
+
+import { getCategoriasError, getCategoriaByIdError, insertCategoria, updateCategoria, deleteCategoria  }from '../controllers/categoriasCRUD';
+
+import { getArticulosError, getArticuloByIdError, insertArticulo, updateArticulo, deleteArticulo  }from '../controllers/articulosCRUD';
+
+import { getReciboMercanciasError, getReciboMercanciaByIdError, insertReciboMercancia }from '../controllers/reciboMercanciasCRUD';
+
+import { getKardexError, getKardexByIdError }from '../controllers/kardexCRUD';
+
+import { getUsuariosError, getUsuarioByIdError, insertUsuario, updateUsuario, deleteUsuario  }from '../controllers/usuariosCRUD';
+
+import { getPermisosError, getPermisoByIdError, insertPermiso, updatePermiso, deletePermiso  }from '../controllers/permisosCRUD';
+
+import { getUsuarioPermisosError, getUsuarioPermisoByIdError, insertUsuarioPermiso, updateUsuarioPermiso, deleteUsuarioPermiso  }from '../controllers/usuarioPermisosCRUD';
+
+import { insertEncabezadoVenta }from '../controllers/encabezadoVentaCRUD';
+
+import { insertDetalleVenta }from '../controllers/detalleVentaCRUD';
 
 
 
@@ -31,51 +45,90 @@ getReciboMercanciaByIdError, getKardexByIdError,getUsuarioByIdError, getPermisoB
  * ******************************************************************************************/
 
 router.get('/clientes', getClientesError);
+
 router.get('/proveedores', getProveedoresError);
+
 router.get('/marcas', getMarcasError);
+
 router.get('/categorias', getCategoriasError);
+
 router.get('/articulos', getArticulosError);
+
 router.get('/reciboMercancias', getReciboMercanciasError);
+
 router.get('/kardex', getKardexError);
+
 router.get('/usuarios', getUsuariosError);
+
 router.get('/permisos', getPermisosError);
+
 router.get('/usuarioPermisos', getUsuarioPermisosError);
+
 
 //consultas por id .
 
 router.get('/clientes/:id', getClienteByIdError);
+
 router.get('/proveedores/:id', getProveedorByIdError);
+
 router.get('/marcas/:id', getMarcaByIdError);
+
 router.get('/categorias/:id', getCategoriaByIdError);
+
 router.get('/articulos/:id', getArticuloByIdError);
+
 router.get('/reciboMercancias/:id', getReciboMercanciaByIdError);
+
 router.get('/kardex/:id', getKardexByIdError);
+
 router.get('/Usuarios/:id', getUsuarioByIdError);
+
 router.get('/permisos/:id', getPermisoByIdError);
+
 router.get('/usuarioPermisos/:id', getUsuarioPermisoByIdError);
+
 
 // insertar datos
 router.post('/clientes', insertCliente);
+
 router.post('/proveedores', insertProveedor);
+
 router.post('/marcas', insertMarca);
+
 router.post('/categorias', insertCategoria);
+
 router.post('/articulos', insertArticulo);
+
 router.post('/reciboMercancias', insertReciboMercancia);
+
 router.post('/encabezadoVenta', insertEncabezadoVenta);
+
 router.post('/detalleVenta', insertDetalleVenta);
+
 router.post('/usuarios', insertUsuario);
+
 router.post('/permisos', insertPermiso);
+
 router.post('/usuarioPermisos', insertUsuarioPermiso);
+
+
 
 
 //actualizar datos
 router.put('/clientes/:id', updateCliente);
+
 router.put('/proveedores/:id', updateProveedor);
+
 router.put('/marcas/:id', updateMarca);
+
 router.put('/categorias/:id', updateCategoria);
+
 router.put('/articulos/:id', updateArticulo);
+
 router.put('/usuarios/:id', updateUsuario);
+
 router.put('/permisos/:id', updatePermiso);
+
 router.put('/usuarioPermisos/:id', updateUsuarioPermiso);
 
 
@@ -83,11 +136,19 @@ router.put('/usuarioPermisos/:id', updateUsuarioPermiso);
 //eliminar registros.
 
 router.delete('/clientes/:id', deleteCliente);
+
 router.delete('/proveedores/:id', deleteProveedor);
+
+router.delete('/marcas/:id', deleteMarca);
+
 router.delete('/categorias/:id', deleteCategoria);
+
 router.delete('/articulos/:id', deleteArticulo);
+
 router.delete('/usuarios/:id', deleteUsuario);
+
 router.delete('/permisos/:id', deletePermiso);
+
 router.delete('/usuarioPermisos/:id', deleteUsuarioPermiso);
